Update SessionManager.swift
Replaced `waitUntilConnected` with `connectionStatus`. Instead of watching `$isConnected`, it listens to `wsPublisher.publisher` and will map incoming values to a `Bool` based on if it receives a `.connected` or a `.disconnected`. If it receives something else, it replaces it with `isConnected`. `$isConnected` is being switched to a non-Published computed property in another commit.