Update SessionManager.swift
Added a `tryFilter` block that checks every event that comes through the stream to see if it's a `.disconnected` event. If it is, it throws a `Errors.disconnected` error. Otherwise, it decides if the event should be allowed through based on if the publisher is connected to the server at that moment.