Update SessionManager.swift
Wrapped all cases of sending a message followed by a publisher listening for a response in a `Publishers.Zip`. This is just in case the response gets back so quickly that the listening publisher hasn't been created yet. The `Zip` is immediately followed by a `.map(\.1)` to pass along only the response of the response publisher.