Update SessionManager Extensions.swift
Added a `debounce(for:scheduler:)` call. This is because sometimes, coming out of studio mode while the preview scene is different than the program scene, the values of the preview and program scene names and the state of studio mode trigger a quick flurry of events that aren't indicative of the correct answer. It ends with the correct value, so this debounce makes sure that a value isn't allowed through unless another value hasn't been received in the following 0.5 seconds.