Update SessionManager.swift
Refactored all publishers (or at least the ones that are either dynamically created, or used to created slightly altered ones based off of them) to be stored, reference-types. This way, they can safely be accessed and subscribed to by multiple sources without any of the sources missing some of the info.