Update SessionManager Extensions.swift
Due to an issue caused by a specific sequence of events, a `.scan` operator was added in. This is in cooperation with and in place of the original `.combineLatest` transform block. Now, by using `.scan`, the `latestPreview` property is overwritten with a `nil` value if it was the last value output, instead of remembering the previous preview scene (which would no longer be correct). The issue was caused if OBS did this sequence:

Studio Mode on > Studio Mode off > Change Scenes (new scene is different from Preview Scene from when Studio Mode was on) > Studio Mode on.